requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
aiohttp